Move serialize_object() & deserialize_object() to utilities.serialization
